WIP: Refactoring code and creating global flag for clipping
