+++ src/lib.js
@@ -36,7 +36,7 @@ import {
   fillBox,
   fillTextLines,
 } from './drawing/index.js';
-import { clip, clipToBox } from './util/clip.js';
+import { clip, clipToBox, clipBoxToDisplayedArea } from './util/clip.js';
 import debounce from './util/debounce';
 import deepmerge from './util/deepmerge';
 import getDefault from './util/getDefault';
@@ -158,6 +158,7 @@ export const lib = {
   'util/convertToVector3': convertToVector3,
   'util/clip': clip,
   'util/clipToBox': clipToBox,
+  'util/clipBoxToDisplayedArea': clipBoxToDisplayedArea,
   'util/debounce': debounce,
   'util/deepmerge': deepmerge,
   'util/getDefault': getDefault,
