+++ src/store/index.js
@@ -26,6 +26,7 @@ export const state = {
   deleteIfHandleOutsideImage: true,
   preventHandleOutsideDisplayedArea: false,
   preventHandleOutsideImage: false,
+  preventTextBoxOutsideDisplayedArea: false,
   // Average pixel width of index finger is 45-57 pixels
   // https://www.smashingmagazine.com/2012/02/finger-friendly-design-ideal-mobile-touchscreen-target-sizes/
   handleTouchOffset: { x: 0, y: -57 },
